Add Meal + Total Macros

+++ src/models/ingredients/Macros.java
@@ -35,7 +35,7 @@ public class Macros {
         return fats;
     }
 
-    public void setFats(int fats) {
+    public void setFats(double fats) {
         this.fats = fats;
     }
